pip_requirements.txt
django >= 1.7, < 1.8
djangorestframework >= 3.1, < 4.0
django-blog-zinnia
pillow
dc-django-base
django-testing-base

